app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
paho-mqtt==2.1.0
apscheduler==3.11.0
httpx==0.28.1
